refactor: Rename frequency mod to rate
